Complete solution to problem 3.

+++ ProjectEuler/EvenFibonacci.cpp
@@ -7,16 +7,16 @@ using namespace std;
 using namespace EulerLibrary::Problems;
 using namespace EulerLibrary;
 
-const long kLimit = 4000000;
+const long long kLimit = 4000000;
 
 string EvenFibonacci::Compute(void)
 {
 	MathLibrary* mathLibrary = new MathLibrary();
 
-    long result = 2;
-    long term1 = 1;
-    long term2 = 2;
-    long sum;
+    long long result = 2;
+    long long term1 = 1;
+    long long term2 = 2;
+    long long sum;
 
     sum = term1 + term2;
     while (sum < kLimit)
